% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/face-package.R
\docType{package}
\name{face-package}
\alias{face-package}
\title{Face Recognition using Principal Component Analysis Package}
\description{
R package for face recognition using
principal componanent analysis.
}
\author{
Al-Ahmadgaid B. Asaad
}

