% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fpr-package.R
\docType{package}
\name{fpr-package}
\alias{fpr-package}
\title{Face Recognition using Principal Component Analysis Package}
\description{
R package for face and pattern recognition using techniques such as
principal componanent analysis and neural networks.
}
\author{
Al-Ahmadgaid B. Asaad
}

