% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fr-package.R
\docType{package}
\name{fr-package}
\alias{fr-package}
\title{Face Recognition using Principal Component Analysis Package}
\description{
This package uses principal component analysis for face recognition.
}
\author{
Al-Ahmadgaid B. Asaad
}

