\name{face-package}
\alias{face-package}
\alias{face}
\docType{package}
\title{face}
\description{
Fast Covariance Estimation for Sparse Functional Data
}
\details{
\tabular{ll}{
Package: \tab face\cr
Type: \tab Package\cr
Version: \tab 0.1-2\cr
Date: \tab 2016-02-26\cr
License: GPL-3
}
}
\author{
Luo Xiao, Cai Li, Will Checkley and Ciprian Crainiceanu

Maintainer: Luo Xiao <lxiao5@ncsu.edu>
}
\references{
Luo Xiao, Cai Li,  Will Checkley and Ciprian Crainiceanu, Fast covariance estimation
for sparse functional data, manuscript.
}
\keyword{ package }

