\name{face-package}
\alias{face-package}
\alias{face}
\docType{package}
\title{face}
\description{
Fast Covariance Estimation for Sparse Functional Data
}
\details{
\tabular{ll}{
Package: \tab face\cr
Type: \tab Package\cr
Version: \tab 0.1-3\cr
Date: \tab 2017-01-02\cr
License: GPL-3
}
}
\author{
Luo Xiao, Cai Li, William Checkley and Ciprian Crainiceanu

Maintainer: Luo Xiao <lxiao5@ncsu.edu>
}
\references{
Luo Xiao, Cai Li, William Checkley and Ciprian Crainiceanu, Fast covariance estimation
for sparse functional data, available at \url{https://arxiv.org/abs/1603.05758}.
}
\keyword{ package }

