\name{face-package}
\alias{face-package}
\alias{face}
\docType{package}
\title{face}
\description{
Fast Covariance Estimation for Sparse Functional Data
}
\details{
\tabular{ll}{
Package: \tab face\cr
Type: \tab Package\cr
Version: \tab 0.1-7\cr
Date: \tab 2022-07-18\cr
License: GPL-3
}
}
\author{
Luo Xiao, Cai Li, William Checkley and Ciprian Crainiceanu

Maintainer: Cai Li <cai.li.stats@gmail.com>
}
\references{
Luo Xiao, Cai Li, William Checkley and Ciprian Crainiceanu, Fast covariance estimation
for sparse functional data, Stat. Comput., \doi{10.1007/s11222-017-9744-8}.
}
\keyword{ package }

